Refactor - moved controllers into folder
